+++ price-feeder/oracle/oracle.go
@@ -69,6 +69,7 @@ type Oracle struct {
 	mtx             sync.RWMutex
 	lastPriceSyncTS time.Time
 	prices          map[string]sdk.Dec
+	paramCache      ParamCache
 }
 
 func New(
@@ -99,6 +100,7 @@ func New(
 		previousPrevote: nil,
 		providerTimeout: providerTimeout,
 		deviations:      deviations,
+		paramCache:      ParamCache{},
 		endpoints:       endpoints,
 	}
 }
@@ -166,7 +168,7 @@ func (o *Oracle) GetPrices() map[string]sdk.Dec {
 // to determine prices. If candles are not available, uses the most recent prices
 // with VWAP. Warns the the user of any missing prices, and filters out any faulty
 // providers which do not report prices or candles within 2𝜎 of the others.
-func (o *Oracle) SetPrices(ctx context.Context, acceptList oracletypes.DenomList) error {
+func (o *Oracle) SetPrices(ctx context.Context) error {
 	g := new(errgroup.Group)
 	mtx := new(sync.Mutex)
 	providerPrices := make(provider.AggregatedProviderPrices)
@@ -356,6 +358,24 @@ func SetProviderTickerPricesAndCandles(
 	return pricesOk || candlesOk
 }
 
+// GetParamCache returns the last updated parameters of the x/oracle module
+// if the current ParamCache is outdated, we will query it again.
+func (o *Oracle) GetParamCache(currentBlockHeigh int64) (oracletypes.Params, error) {
+	if !o.paramCache.IsOutdated(currentBlockHeigh) {
+		return *o.paramCache.params, nil
+	}
+
+	params, err := o.GetParams()
+	if err != nil {
+		return oracletypes.Params{}, err
+	}
+
+	o.checkAcceptList(params)
+
+	o.paramCache.Update(currentBlockHeigh, params)
+	return params, nil
+}
+
 // GetParams returns the current on-chain parameters of the x/oracle module.
 func (o *Oracle) GetParams() (oracletypes.Params, error) {
 	grpcConn, err := grpc.Dial(
@@ -460,15 +480,6 @@ func (o *Oracle) tick(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
-	oracleParams, err := o.GetParams()
-	if err != nil {
-		return err
-	}
-	if err := o.SetPrices(ctx, oracleParams.AcceptList); err != nil {
-		return err
-	}
-
-	o.checkAcceptList(oracleParams)
 
 	blockHeight, err := rpcclient.GetChainHeight(clientCtx)
 	if err != nil {
@@ -476,6 +487,15 @@ func (o *Oracle) tick(ctx context.Context) error {
 	}
 	if blockHeight < 1 {
 		return fmt.Errorf("expected positive block height")
+	}
+
+	oracleParams, err := o.GetParamCache(blockHeight)
+	if err != nil {
+		return err
+	}
+
+	if err := o.SetPrices(ctx); err != nil {
+		return err
 	}
 
 	// Get oracle vote period, next block height, current vote period, and index